+++ CHANGELOG.md
@@ -5,9 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project follows [Semantic Versioning](https://semver.org/).
 
-## [Unreleased]
+## [1.0.0] 2026.09.17 (In Progress)
+
+### Changed
+
+    - Converted wiki to JavaScript based blog.
+
+## [0.2] 2025.09.25
 
 ### Added
+
+    - Created file with Tiddlywiki v5.2.8
 
 ### Changed